Test Refactorings
